# api_docs.yaml
openapi: 3.0.0
info:
  title: Business Management System API
  version: 1.0.0
  description: API documentation for the Business Management System
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.example.com/api
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /inventory/adjust:
    post:
      summary: Adjust product stock
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                product_id:
                  type: integer
                quantity:
                  type: integer
                type:
                  type: string
                  enum: [receive, adjust, return]
                notes:
                  type: string
      responses:
        '200':
          description: Stock adjusted successfully
        '400':
          description: Invalid request
        '404':
          description: Product not found
...